reorganize crate tests
