+++ tests/compile-pass/redefinition/core.rs
@@ -5,18 +5,10 @@ use bitflags::bitflags;
 mod core {}
 mod _core {}
 
-#[allow(unused_macros)]
-macro_rules! stringify {
-    ($($t:tt)*) => { "..." };
-}
-
 bitflags! {
     struct Test: u8 {
         const A = 1;
     }
 }
 
-#[test]
-fn stringify() {
-    assert_eq!(format!("{:?}", Test::A), "A");
-}
+fn main() {}
